Golang backend

+++ src/store/store.ts
@@ -1,10 +1,12 @@
 /* eslint-disable no-console */
-import dayjs from "dayjs";
 import { action, computed, createStore, thunk } from "easy-peasy";
 import _ from "lodash";
 import authService from "../services/auth.service";
 import itemsService from "../services/items.service";
 import listsService from "../services/lists.service";
+import { ApiError } from "../types/API";
+import { Item } from "../types/items";
+import { SSOError } from "../types/SSO";
 import { StoreModel } from "./models";
 
 export const store = createStore<StoreModel>({
@@ -12,19 +14,18 @@ export const store = createStore<StoreModel>({
     const ssoError = await actions.auth.fetch();
     if (ssoError) {
       console.log({ ssoError });
-      return false;
+      return [false, ssoError] as [boolean, SSOError];
     }
 
-    const errors = await Promise.all([
+    const errors: (ApiError | null)[] = await Promise.all([
       actions.items.fetch(),
       actions.lists.fetch(),
     ]);
 
     if (errors.some((x) => x !== null)) {
-      console.log({ errors });
-      return false;
+      return [false, errors] as [boolean, (ApiError | null)[]];
     }
-    return true;
+    return [true, null] as [boolean, null];
   }),
   core: {
     loaded: false,
@@ -96,12 +97,12 @@ export const store = createStore<StoreModel>({
     addItem: thunk(async (actions, payload) => {
       const [item, error] = await itemsService.addItem(payload);
       if (error) {
-        return error;
+        return [null, error] as [null, ApiError];
       }
       if (item) {
         actions.append(item);
       }
-      return null;
+      return [item, null] as [Item, null];
     }),
     removeItem: thunk(async (actions, payload) => {
       const error = await itemsService.deleteItem(payload.id);
@@ -114,8 +115,11 @@ export const store = createStore<StoreModel>({
   },
   lists: {
     lists: [],
+    defaultListStore: null,
     defaultList: computed((state) => {
-      const defaultList = state.lists.find((x) => x.default);
+      const defaultList = state.lists.find(
+        (x) => x.id === state.defaultListStore?.listId
+      );
       return defaultList || null;
     }),
     setAll: action((state, payload) => {
@@ -123,23 +127,40 @@ export const store = createStore<StoreModel>({
     }),
     set: action((state, payload) => {
       const index = state.lists.findIndex((x) => x.id === payload.id);
-      if (payload.default) {
-        state.lists.forEach((list) => {
-          list.default = false;
-        });
-      }
       if (index !== -1) {
         state.lists[index] = payload;
       } else {
         state.lists.push(payload);
       }
     }),
-    append: action((state, payload) => {
-      if (payload.default) {
-        state.lists.forEach((list) => {
-          list.default = false;
-        });
+    setItem: action((state, payload) => {
+      const listIndex = state.lists.findIndex((x) => x.id === payload.list.id);
+      if (listIndex !== -1) {
+        const list = state.lists[listIndex];
+        const itemIndex = list.items.findIndex(
+          (x) => x.id === payload.listItem.id
+        );
+        if (itemIndex !== -1) {
+          list.items[itemIndex] = payload.listItem;
+        }
       }
+    }),
+    removeItem: action((state, payload) => {
+      const listIndex = state.lists.findIndex((x) => x.id === payload.list.id);
+      if (listIndex !== -1) {
+        const list = state.lists[listIndex];
+        const itemIndex = list.items.findIndex(
+          (x) => x.id === payload.listItem.id
+        );
+        if (itemIndex !== -1) {
+          list.items.splice(itemIndex, 1);
+        }
+      }
+    }),
+    setDefaultList: action((state, payload) => {
+      state.defaultListStore = payload;
+    }),
+    append: action((state, payload) => {
       state.lists.push(payload);
     }),
     removeById: action((state, payload) => {
@@ -155,6 +176,16 @@ export const store = createStore<StoreModel>({
       }
       if (lists) {
         actions.setAll(lists);
+      }
+      const [
+        defaultList,
+        defaultListError,
+      ] = await listsService.getDefaultList();
+      if (defaultListError) {
+        return defaultListError;
+      }
+      if (defaultList) {
+        actions.setDefaultList(defaultList);
       }
       return null;
     }),
@@ -175,6 +206,16 @@ export const store = createStore<StoreModel>({
       }
       if (list) {
         actions.set(list);
+      }
+      return null;
+    }),
+    updateDefaultList: thunk(async (actions, payload) => {
+      const [defaultList, error] = await listsService.setDefaultList(payload);
+      if (error) {
+        return error;
+      }
+      if (defaultList) {
+        actions.setDefaultList(defaultList);
       }
       return null;
     }),
@@ -199,6 +240,9 @@ export const store = createStore<StoreModel>({
       if (listItem) {
         listItem.item = payload.item;
         const clonedList = _.cloneDeep(payload.list);
+        if (_.isNil(clonedList.items)) {
+          clonedList.items = [];
+        }
         clonedList.items.push(listItem);
         actions.set(clonedList);
       }
@@ -206,23 +250,29 @@ export const store = createStore<StoreModel>({
     }),
 
     removeFromList: thunk(async (actions, payload) => {
-      const error = await listsService.removeFromList({
-        listId: payload.list.id,
-        listItemId: payload.listItem.id,
-      });
-      if (error) {
-        return error;
-      }
+      if (!payload.listItem.crossed) {
+        const listItemClone = _.cloneDeep(payload.listItem);
+        listItemClone.crossed = true;
+        const [updatedListItem, error] = await listsService.updateListItem(
+          listItemClone
+        );
+        if (error) {
+          return error;
+        }
 
-      const clonedList = _.cloneDeep(payload.list);
-      const itemIndex = clonedList.items.findIndex(
-        (x) => x.id === payload.listItem.id
-      );
-      if (itemIndex !== -1) {
-        clonedList.items.splice(itemIndex, 1);
+        if (updatedListItem) {
+          actions.setItem({ list: payload.list, listItem: updatedListItem });
+        }
+      } else {
+        const error = await listsService.removeFromList({
+          listId: payload.list.id,
+          listItemId: payload.listItem.id,
+        });
+        if (error) {
+          return error;
+        }
+        actions.removeItem({ list: payload.list, listItem: payload.listItem });
       }
-      actions.set(clonedList);
-
       return null;
     }),
   },
